refactor: 패키지 분류 세분화
