+++ src/main/java/cart/controller/ProductController.java
@@ -1,5 +1,6 @@
 package cart.controller;
 
+import cart.controller.dto.ProductRequest;
 import cart.dao.ProductDao;
 import java.util.NoSuchElementException;
 import javax.validation.Valid;
